requirements.txt
streamlit
pandas
yfinance
alpha-vantage
finnhub-python
python-dotenv
plotly
scikit-learn
pandas-ta
requests
flask
scipy
xgboost
tensorflow
shap
textblob
newsapi-python
pytest
numpy
